requirements.txt
boto3
pdfplumber
flask
python-dotenv
gunicorn
